requirements.txt
asyncpg>=0.29.0
discord.py>=2.4.0
tinytag>=1.10.1
wavelink>=3.4.1